Make windows tests happy.
